+++ src/helpers/node-text-utils.ts
@@ -10,14 +10,36 @@ type NodeHandler = (nodes: Node[]) => void;
  *
  * @param selector CSS selector to find nodes by
  * @param handler handler to pass nodes to
+ * @param parentSelector CSS selector to find parent nodes by
  */
 export const handleExistingNodes = (
     selector: string,
     handler: NodeHandler,
+    parentSelector?: string,
 ): void => {
-    const nodeList = document.querySelectorAll(selector);
-    const nodes = nodeListToArray(nodeList);
-    handler(nodes);
+    /**
+     * Processes nodes within a given parent element based on the provided selector.
+     * If the selector is '#text', it will filter and handle text nodes.
+     * Otherwise, it will handle elements that match the provided selector.
+     *
+     * @param parent Parent node in which to search for nodes.
+     */
+    const processNodes = (parent: ParentNode) => {
+        // If the selector is '#text', filter and handle text nodes.
+        if (selector === '#text') {
+            const textNodes = nodeListToArray(parent.childNodes)
+                .filter((node) => node.nodeType === Node.TEXT_NODE);
+            handler(textNodes);
+        } else {
+            // Handle elements that match the provided selector
+            const nodes = nodeListToArray(parent.querySelectorAll(selector));
+            handler(nodes);
+        }
+    };
+    // If a parent selector is provided, process nodes within each parent element.
+    // If not, process nodes within the document.
+    const parents = parentSelector ? document.querySelectorAll(parentSelector) : [document];
+    parents.forEach((parent) => processNodes(parent));
 };
 
 /**
@@ -25,13 +47,23 @@ export const handleExistingNodes = (
  *
  * @param mutations mutations to find eligible nodes in
  * @param handler handler to pass eligible nodes to
+ * @param selector CSS selector to find nodes by
+ * @param parentSelector CSS selector to find parent nodes by
  */
 export const handleMutations = (
     mutations: MutationRecord[],
     handler: NodeHandler,
+    selector?: string,
+    parentSelector?: string,
 ): void => {
     const addedNodes = getAddedNodes(mutations);
-    handler(addedNodes);
+    if (selector && parentSelector) {
+        addedNodes.forEach(() => {
+            handleExistingNodes(selector, handler, parentSelector);
+        });
+    } else {
+        handler(addedNodes);
+    }
 };
 
 /**